+++ ApiBlibliotecaSimples/Entities/ItemVenda.cs
@@ -6,11 +6,11 @@ public class ItemVenda
     public int Quantidade { get; private set; }
     public decimal PrecoUnitario { get; private set; }
 
-    public long ExemplarId { get; private set; }
+    public int ExemplarId { get; private set; }
     public ExemplarLivro Exemplar { get; private set; }
 
     public ItemVenda() { }
-    public ItemVenda(int quantidade, decimal precoUnitario, long exemplarId)
+    public ItemVenda(int quantidade, decimal precoUnitario, int exemplarId)
     {
         Quantidade = quantidade;
         PrecoUnitario = precoUnitario;
